cmd: use new architecture client by default except `cdc server`
