+++ cmd/cdc/server/server.go
@@ -16,6 +16,7 @@ package server
 import (
 	"context"
 	"os"
+	"slices"
 	"strings"
 
 	"github.com/fatih/color"
@@ -26,6 +27,7 @@ import (
 	"github.com/pingcap/ticdc/pkg/logger"
 	"github.com/pingcap/ticdc/pkg/version"
 	"github.com/pingcap/ticdc/server"
+	tiflowServer "github.com/pingcap/tiflow/pkg/cmd/server"
 	"github.com/pingcap/tiflow/pkg/security"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
@@ -54,6 +56,7 @@ func newOptions() *options {
 // addFlags receives a *cobra.Command reference and binds
 // flags related to template printing to it.
 func (o *options) addFlags(cmd *cobra.Command) {
+	cmd.Flags().BoolVarP(&o.serverConfig.Newarch, "newarch", "x", o.serverConfig.Newarch, "Run the new architecture of TiCDC server")
 	cmd.Flags().StringVar(&o.serverConfig.ClusterID, "cluster-id", "default", "Set cdc cluster id")
 	cmd.Flags().StringVar(&o.serverConfig.Addr, "addr", o.serverConfig.Addr, "Set the listening address")
 	cmd.Flags().StringVar(&o.serverConfig.AdvertiseAddr, "advertise-addr", o.serverConfig.AdvertiseAddr, "Set the advertise listening address for client communication")
@@ -206,6 +209,57 @@ func (o *options) getCredential() *security.Credential {
 	}
 }
 
+func parseConfigFlagFromOSArgs() string {
+	var serverConfigFilePath string
+	for i, arg := range os.Args[1:] {
+		if strings.HasPrefix(arg, "--config=") {
+			serverConfigFilePath = strings.SplitN(arg, "=", 2)[1]
+		} else if arg == "--config" && i+2 < len(os.Args) {
+			serverConfigFilePath = os.Args[i+2]
+		}
+	}
+
+	// If the command is `cdc cli changefeed`, means it's not a server config file.
+	if slices.Contains(os.Args, "cli") && slices.Contains(os.Args, "changefeed") {
+		serverConfigFilePath = ""
+	}
+
+	return serverConfigFilePath
+}
+
+func isNewArchEnabledByConfig(serverConfigFilePath string) bool {
+	cfg := config.GetDefaultServerConfig()
+	if len(serverConfigFilePath) > 0 {
+		// strict decode config file, but ignore debug item
+		if err := util.StrictDecodeFile(serverConfigFilePath, "TiCDC server", cfg, config.DebugConfigurationItem); err != nil {
+			log.Error("failed to parse server configuration, please check the config file for errors and try again.", zap.Error(err))
+			return false
+		}
+	}
+
+	return cfg.Newarch
+}
+
+func isNewArchEnabled(o *options) bool {
+	newarch := o.serverConfig.Newarch
+	if newarch {
+		log.Info("Set newarch from command line")
+	}
+
+	newarch = os.Getenv("TICDC_NEWARCH") == "true"
+	if newarch {
+		log.Info("Set newarch from environment variable")
+	}
+
+	serverConfigFilePath := parseConfigFlagFromOSArgs()
+	newarch = isNewArchEnabledByConfig(serverConfigFilePath)
+	if newarch {
+		log.Info("Set newarch from config file")
+	}
+
+	return newarch
+}
+
 // NewCmdServer creates the `server` command.
 func NewCmdServer() *cobra.Command {
 	o := newOptions()
@@ -215,17 +269,20 @@ func NewCmdServer() *cobra.Command {
 		Short: "Start a TiCDC server server",
 		Args:  cobra.NoArgs,
 		RunE: func(cmd *cobra.Command, args []string) error {
-			err := o.complete(cmd)
-			if err != nil {
-				return err
+			if isNewArchEnabled(o) {
+				err := o.complete(cmd)
+				if err != nil {
+					return err
+				}
+				err = o.validate()
+				if err != nil {
+					return err
+				}
+				err = o.run(cmd)
+				cobra.CheckErr(err)
+				return nil
 			}
-			err = o.validate()
-			if err != nil {
-				return err
-			}
-			err = o.run(cmd)
-			cobra.CheckErr(err)
-			return nil
+			return tiflowServer.Run(o, cmd)
 		},
 	}
